AI-agent-for-bank-/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
google-generativeai
langchain-google-genai
langchain
pypdf
faiss-cpu
langchain-community
unstructured
libmagic
langchain-ollama
chromadb
Pillow
SpeechRecognition
PyAudio